chore(debt): sync final baseline gates
