+++ docs/performance-baseline.md
@@ -1,6 +1,6 @@
 # 性能优化与热点基线
 
-日期：2026-05-18
+日期：2026-05-21
 
 ## 1. 构建真值
 
@@ -11,7 +11,7 @@ npm run build:dev
 npm run audit:build:report
 ```
 
-2026-05-18 stabilization 复核使用 fresh build：
+2026-05-21 M7 final baseline sync 复核使用 fresh build：
 
 ```bash
 npm run clean
@@ -22,10 +22,10 @@ npm run audit:build:report
 当前真值：
 
 - `build/dist/content/index.js`: `561 B`
-- `build/dist/content/runtime.js`: `56.0 KB`
+- `build/dist/content/runtime.js`: `53.3 KB`（raw `54,554` bytes；距离 `57,600` raw-byte stop gate 还有 `3,046` bytes）
 - `build/dist/options/index.js`: `997 B`
 - `build/dist/onboarding/index.js`: `12.3 KB`
-- 总 chunk 数：`98`
+- 总 chunk 数：`102`
 
 当前 shared chunk Top 3（`chunk-*`，按 `tools/report-build-splitting.mjs` 口径）：
 
@@ -35,9 +35,10 @@ npm run audit:build:report
 
 当前大型 runtime/lazy chunks（不等同于 shared Top 3 预算口径）：
 
-- `chunks/runtimeEntry-*.js`: `262.5 KB`
-- `chunks/videoSessionControllers-*.js`: `71.0 KB`
-- `chunks/videoLazyRuntime-*.js`: `37.0 KB`
+- `chunks/runtimeEntry-*.js`: `276.8 KB`
+- `chunks/videoSessionControllers-*.js`: `70.9 KB`
+- `chunks/qps-ploc-*.js`: `57.1 KB`
+- `chunks/videoLazyRuntime-*.js`: `37.8 KB`
 
 当前重点功能 chunk：
 
@@ -47,7 +48,7 @@ npm run audit:build:report
 
 当前 `audit:build:report` 预算口径：
 
-- `content/runtime.js <= 56 KB`
+- `content/runtime.js <= 56 KB`（workspace stop gate 另按 raw `57,600` bytes 执行）
 - `options/index.js <= 107 KB`
 - `onboarding/index.js <= 20 KB`
 - 任一 chunk `<= 650 KB`
@@ -72,10 +73,14 @@ npm run audit:performance:report
 - `src/content/video/videoSessionRuntime.ts`: `395` 行
 - `src/options/components/sections/RestSectionView.ts`: `260` 行
 - `src/ui/domains/privacy/PrivacySettingsView.ts`: `255` 行
-- `src/options/components/sections/UsageDashboardSection.ts`: `172` 行
-- `src/ui/domains/yaml-config/yamlConfigTableRenderer.ts`: `233` 行
-- `src/ui/domains/yaml-config/yamlConfigTableStateModel.ts`: `183` 行
 - `src/content/reader/utils/markdownBuilder.ts`: `288` 行
+- `src/content/extractors/articleExtractor.ts`: `222` 行
+- `src/options/components/sections/FragmentSectionView.ts`: `204` 行
+- `src/options/state/optionsStore.ts`: `194` 行
+- `src/ui/domains/yaml-config/yamlConfigTableStateModel.ts`: `183` 行
+- `src/content/video/platforms/bilibiliPlatformAdapter.ts`: `178` 行
+- `src/options/components/sections/UsageDashboardSection.ts`: `173` 行
+- `src/content/index.ts`: `154` 行
 - `src/options/state/StateManager.ts`: `deepClone=0`, `JSON.stringify=0`
 - `src/options/state/optionsStore.ts`: `deepClone=0`, `JSON.stringify=0`
 
@@ -84,7 +89,7 @@ npm run audit:performance:report
 - `videoSessionRuntime` 当前为 `395` 行，仍需作为 runtime hotspot 观察项
 - `PrivacySettingsView`、`UsageDashboardSection` 已达到 stretch 目标
 - `RestSectionView` 已落到本轮最低目标内
-- YAML config table state/model 已拆分到当前热点口径内
+- YAML config table renderer/state/model 已拆分到当前热点口径内；`yamlConfigTableRenderer.ts` 当前为 `67` 行
 - 本轮新增子模块已按热点治理范围压到目标值，避免用拆分制造新的超限热点
 - Options 低频 diagnostics 已从主入口拆出
 - `YamlConfigView` 已改为 view/controller 双层 lazy；2026-05-18 build report 不再产出旧 `yaml-config-*` chunk 名称
@@ -108,6 +113,6 @@ npm run audit:performance:report
 
 ## 4. 债务备注
 
-- `tools/baselines/lint-warnings.json` 基线仍记录历史 warning 债务；checked-in warning guard baseline 为 `322` 条，2026-05-21 fresh warning count 为 `274` 条，不代表 lint 债务已经清零。`lint:warnings-report` 会重写该 baseline；baseline 下调留到 M7 final baseline sync。
+- `tools/baselines/lint-warnings.json` 基线仍记录历史 warning 债务；2026-05-21 M7 已将 checked-in warning guard baseline 从 `322` 下调到 `266` 条。`lint:warnings-report` 仍会重写该 baseline，只能在有意同步 warning truth 时运行。
 - Firefox build path 已在 2026-05-18 stabilization 中通过 `npm run build:firefox`；Firefox browser smoke 仍不是本轮强制浏览器收口范围。
-- 当前本地执行环境 Node.js 为 `v23.9.0`，高于 `package.json` 声明的 `>=20 <21`，验证命令通过但仍应回到 Node `20.x` 作为发布环境。
+- 当前本地执行环境 Node.js 为 `v23.9.0`，高于 `package.json` 声明的 `>=20 <21`；M7 本地验证通过但最终发布证据仍必须回到 Node `20.x`。
