+++ src/options/app/productionStitchStorageController.ts
@@ -1,35 +1,16 @@
-import type { IMessagingRepository } from '@shared/repositories';
-import type { CompleteOptions } from '@shared/types/options';
-import type { ConnectionTestResult } from '@shared/types/connection';
-import type { VaultRouterConfig } from '@shared/types/vault';
-import type { PreviewContent, PreviewStoreState } from '@options/stitch/types';
 import { createProductionStitchStorageFeedback } from './productionStitchStorageFeedback';
 import { createProductionStitchStorageLoad } from './productionStitchStorageLoad';
 import { createProductionStitchStorageSave } from './productionStitchStorageSave';
 import { createProductionStitchStorageSubscriptions } from './productionStitchStorageSubscriptions';
+import type {
+  ProductionStitchStorageController,
+  ProductionStitchStorageControllerOptions
+} from './productionStitchStorageTypes';
 
-export interface ProductionStitchStorageControllerOptions {
-  getConnectionNotice(): PreviewContent['storage']['connectionNotice'] | undefined;
-  getDraft(): CompleteOptions;
-  getMessagingRepository(): Pick<IMessagingRepository, 'send' | 'onMessage'>;
-  getState(): PreviewStoreState;
-  setConnectionNotice(notice: PreviewContent['storage']['connectionNotice'] | undefined): void;
-  refreshAppData(): void;
-  render(): void;
-  scheduleDraftSave(): void;
-}
-
-export interface ProductionStitchStorageController {
-  activateVaultLocalFolder(index: number): Promise<void>;
-  applyConnectionNotice(result: ConnectionTestResult): void;
-  chooseVaultLocalFolder(index: number): Promise<void>;
-  clearVaultLocalFolder(index: number): void;
-  ensureVaultRouter(): VaultRouterConfig;
-  runVaultListConnectionTest(): Promise<ConnectionTestResult>;
-  syncDefaultVaultFromRest(): void;
-  syncRoutingRulesToDraft(): void;
-  updateVaultField(index: number, field: string, value: unknown): void;
-}
+export type {
+  ProductionStitchStorageController,
+  ProductionStitchStorageControllerOptions
+} from './productionStitchStorageTypes';
 
 export function createProductionStitchStorageController(
   options: ProductionStitchStorageControllerOptions
@@ -40,14 +21,14 @@ export function createProductionStitchStorageController(
   const feedback = createProductionStitchStorageFeedback(options, load);
 
   return {
-    activateVaultLocalFolder: subscriptions.activateVaultLocalFolder,
-    applyConnectionNotice: feedback.applyConnectionNotice,
-    chooseVaultLocalFolder: subscriptions.chooseVaultLocalFolder,
-    clearVaultLocalFolder: subscriptions.clearVaultLocalFolder,
-    ensureVaultRouter: load.ensureVaultRouter,
-    runVaultListConnectionTest: feedback.runVaultListConnectionTest,
-    syncDefaultVaultFromRest: load.syncDefaultVaultFromRest,
-    syncRoutingRulesToDraft: save.syncRoutingRulesToDraft,
-    updateVaultField: save.updateVaultField
+    activateVaultLocalFolder: (index) => subscriptions.activateVaultLocalFolder(index),
+    applyConnectionNotice: (result) => feedback.applyConnectionNotice(result),
+    chooseVaultLocalFolder: (index) => subscriptions.chooseVaultLocalFolder(index),
+    clearVaultLocalFolder: (index) => subscriptions.clearVaultLocalFolder(index),
+    ensureVaultRouter: () => load.ensureVaultRouter(),
+    runVaultListConnectionTest: () => feedback.runVaultListConnectionTest(),
+    syncDefaultVaultFromRest: () => load.syncDefaultVaultFromRest(),
+    syncRoutingRulesToDraft: () => save.syncRoutingRulesToDraft(),
+    updateVaultField: (index, field, value) => save.updateVaultField(index, field, value)
   };
 }
